V2

+++ routes/web.php
@@ -31,45 +31,44 @@ Route::get('/dashboard', function () {
 })->middleware(['auth'])->name('dashboard');
 
 // Routes Dashboard
-Route::get('/dashboard', [DashboardController::class, 'index']);
-Route::get('/profiles', [DashboardController::class, 'profile']);
-Route::get('/profile/edit/{id}', [DashboardController::class, 'editProfile']);
-Route::patch('/profile/edit/{id}', [DashboardController::class, 'updateProfile']);
+Route::get('/dashboard', [DashboardController::class, 'index'])->middleware(['auth']);
+Route::get('/profiles', [DashboardController::class, 'profile'])->middleware(['auth']);
+Route::get('/profile/edit/{id}', [DashboardController::class, 'editProfile'])->middleware(['auth']);
+Route::patch('/profile/edit/{id}', [DashboardController::class, 'updateProfile'])->middleware(['auth']);
 
 // Routes Data Karyawan
-Route::get('/karyawan', [KaryawanController::class, 'index']);
-Route::post('/karyawan', [KaryawanController::class, 'store']);
-Route::get('/karyawan/detail/{id}/', [KaryawanController::class, 'show']);
-Route::get('/karyawan/edit/{id}/', [KaryawanController::class, 'edit']);
-Route::patch('/karyawan/update/{id}', [KaryawanController::class, 'update']);
-Route::delete('/karyawan/delete/{id}', [KaryawanController::class, 'delete']);
+Route::get('/karyawan', [KaryawanController::class, 'index'])->middleware(['auth', 'admin', 'owner']);
+Route::post('/karyawan', [KaryawanController::class, 'store'])->middleware(['auth']);
+Route::get('/karyawan/detail/{id}/', [KaryawanController::class, 'show'])->middleware(['auth', 'admin', 'owner']);
+Route::get('/karyawan/edit/{id}/', [KaryawanController::class, 'edit'])->middleware(['auth', 'admin']);
+Route::patch('/karyawan/update/{id}', [KaryawanController::class, 'update'])->middleware(['auth', 'admin']);
+Route::delete('/karyawan/delete/{id}', [KaryawanController::class, 'delete'])->middleware(['auth', 'admin']);
 
 // Routes Data Jabatan
-Route::get('/jabatan', [JabatanController::class, 'index']);
-Route::get('/jabatan/create', [JabatanController::class, 'create']);
-Route::post('/jabatan', [JabatanController::class, 'store']);
-Route::get('/jabatan/{id}/', [JabatanController::class, 'edit']);
-Route::patch('/jabatan/{id}', [JabatanController::class, 'update']);
-Route::delete('/jabatan/{id}', [JabatanController::class, 'destroy']);
+Route::get('/jabatan', [JabatanController::class, 'index'])->middleware(['auth', 'admin', 'owner']);
+Route::get('/jabatan/create', [JabatanController::class, 'create'])->middleware(['auth', 'admin', 'owner']);
+Route::post('/jabatan', [JabatanController::class, 'store'])->middleware(['auth', 'admin', 'owner']);
+Route::get('/jabatan/{id}/', [JabatanController::class, 'edit'])->middleware(['auth', 'admin', 'owner']);
+Route::patch('/jabatan/{id}', [JabatanController::class, 'update'])->middleware(['auth', 'admin', 'owner']);
+Route::delete('/jabatan/{id}', [JabatanController::class, 'destroy'])->middleware(['auth', 'admin', 'owner']);
 
 // Routes Absensi
-Route::get('/absensi', [AbsensiController::class, 'index']);
-Route::post('/absensi/in', [AbsensiController::class, 'absensiIn']);
-Route::post('/absensi/out', [AbsensiController::class, 'absensiOut']);
+Route::get('/absensi', [AbsensiController::class, 'index'])->middleware(['auth']);
+Route::post('/absensi', [AbsensiController::class, 'absen'])->middleware(['auth']);
 
 // Routes Data Penggajian
-Route::get('/payroll', [PayrollController::class, 'index']);
-Route::get('/payroll/edit/{id}', [PayrollController::class, 'edit']);
-Route::patch('/payroll/update/{id}', [PayrollController::class, 'update']);
-Route::get('/payroll/print/{id}', [PayrollController::class, 'print']);
-Route::get('/exportPDF', [PayrollController::class, 'exportPDF']);
+Route::get('/payroll', [PayrollController::class, 'index'])->middleware(['auth', 'admin', 'owner']);
+Route::get('/payroll/edit/{id}', [PayrollController::class, 'edit'])->middleware(['auth', 'admin', 'owner']);
+Route::patch('/payroll/update/{id}', [PayrollController::class, 'update'])->middleware(['auth', 'admin', 'owner']);
+Route::get('/payroll/print/{id}', [PayrollController::class, 'print'])->middleware(['auth', 'admin', 'owner']);
+Route::get('/exportPDF', [PayrollController::class, 'exportPDF'])->middleware(['auth', 'admin', 'owner']);
 
 // Route Leave Request
-Route::get('/leave-request', [LeaveRequestController::class, 'index']);
-Route::get('/leave-request/create', [LeaveRequestController::class, 'create']);
-Route::post('/leave-request', [LeaveRequestController::class, 'store']);
-Route::patch('/leave-request/approved/{id}/', [LeaveRequestController::class, 'approved']);
-Route::patch('/leave-request/rejected/{id}', [LeaveRequestController::class, 'rejected']);
+Route::get('/leave-request', [LeaveRequestController::class, 'index'])->middleware(['auth']);
+Route::get('/leave-request/create', [LeaveRequestController::class, 'create'])->middleware(['auth']);
+Route::post('/leave-request', [LeaveRequestController::class, 'store'])->middleware(['auth']);
+Route::patch('/leave-request/approved/{id}/', [LeaveRequestController::class, 'approved'])->middleware(['auth']);
+Route::patch('/leave-request/rejected/{id}', [LeaveRequestController::class, 'rejected'])->middleware(['auth']);
 // Group Admin Routes
 
 // Group Owner Routes
